refactor : Parser 도입
