+++ src/main/java/chess/controller/ChessGameController.java
@@ -2,19 +2,14 @@ package chess.controller;
 
 import chess.domain.BoardInitializer;
 import chess.domain.ChessGame;
-import chess.domain.position.Column;
 import chess.domain.position.Position;
-import chess.domain.position.Row;
-import chess.view.ColumnMapper;
+import chess.dto.PositionParser;
+import chess.dto.CommandDto;
 import chess.view.Commend;
-import chess.view.CommandDto;
 import chess.view.InputView;
 import chess.view.OutputView;
-import chess.view.RowMapper;
 
 public class ChessGameController {
-    private static final int INPUT_COLUMN_INDEX = 0;
-    private static final int INPUT_ROW_INDEX = 1;
 
     private final InputView inputView;
     private final OutputView outputView;
@@ -62,13 +57,11 @@ public class ChessGameController {
     }
 
     private void handleMoveCommend(ChessGame chessGame, CommandDto commandDto) {
-        chessGame.handleMove(createPositionByString(commandDto.from()), createPositionByString(commandDto.to()));
+        Position fromPosition = PositionParser.parsing(commandDto.from());
+        Position toPosition = PositionParser.parsing(commandDto.to());
+        chessGame.handleMove(fromPosition, toPosition);
         outputView.printBoard(chessGame.getBoard());
     }
 
-    private Position createPositionByString(String positionValue) {
-        Row row = RowMapper.findByInputValue(positionValue.split("")[INPUT_ROW_INDEX]);
-        Column column = ColumnMapper.findByInputValue(positionValue.split("")[INPUT_COLUMN_INDEX]);
-        return new Position(row, column);
-    }
+
 }
