+++ src/main/java/chess/view/InputView.java
@@ -1,5 +1,6 @@
 package chess.view;
 
+import chess.dto.CommandDto;
 import java.util.Scanner;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
